requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
clip==0.2.0
contourpy==1.1.1
cycler==0.12.1
decord==0.6.0
einops==0.8.0
-e git+https://github.com/VQAssessment/FAST-VQA-and-FasterVQA.git@9f264e436a97b23f5efb477b048bfbff83565032#egg=fastvqa
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.0
fsspec==2024.6.1
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
safetensors==0.4.3
scipy==1.10.1
siti @ file:///home/sunwei/code/VQA/siti
six==1.16.0
sk-video==1.1.10
sympy==1.12.1
timm==1.0.7
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.6
zipp==3.20.2
